Open source DB in read/write mode when copying database (fixes #132)
Even though it is possible to open databases in WAL mode as read-only
since sqlite 3.22 (released in 2018), there seem to be a couple of limitations:
https://sqlite.org/wal.html#readonly

The -wal and -shm are left behind even after the database is closed.
So better let's open it in read/write mode as usual and those files
will get proper treatment - they will be removed when connection is closed.