Fix date/time handling when string values are not the same (fixes #143)
This fixes the issue when applying a diff with date/time values,
but the representation of date/time values as strings is slightly
different even though they refer to the same date/time.

Fixes both the case with update and the case with delete.
Insert does not need fixing as there is no comparison with previous values.